analemma-workflow-os/backend/src/requirements.txt
boto3
botocore
croniter
pytest
moto

google-generativeai
langgraph>=0.0.40
openai
jsonpath-ng
jsonschema
pydantic>=2.7.4
dill==0.3.6
aws-lambda-powertools
PyJWT[crypto]
ijson>=3.2.0  # 스트리밍 JSON 파서 (대용량 S3 결과 처리용)

# AWS X-Ray (aws-lambda-powertools에서 필요)
aws-xray-sdk>=2.12.0

# FastAPI 및 웹 프레임워크
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mangum>=0.17.0
python-multipart>=0.0.6

# LangGraph 및 워크플로우
langgraph-checkpoint-dynamodb>=0.0.1

# 추가 유틸리티
requests>=2.31.0
aiohttp>=3.9.0
urllib3>=1.26.0
